requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.1
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.1.10
django-celery-results==2.5.1
django-ninja==0.22.2
dnspython==2.4.1
kombu==5.3.1
pika==1.3.2
Pillow==10.0.0
prompt-toolkit==3.0.39
pydantic==1.10.12
pymongo==4.4.1
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.2.4
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
psycopg2-binary
redis==4.6.0
pytest_django==4.5.2
